pyproject.toml


[project]
name = "hackathon"
version = "0.1.0"
description = "Mercurious Hackathon Team work"
readme = "README-cloudshell.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "google-adk>=1.0.0",
    "PyYAML>=6.0.2, <7.0.0",                                   # For APIHubToolset.
    "aiosqlite>=0.21.0",                                       # For SQLite database
    "anyio>=4.9.0, <5.0.0",                                    # For MCP Session Manager
    "authlib>=1.5.1, <2.0.0",                                  # For RestAPI Tool
    "click>=8.1.8, <9.0.0",                                    # For CLI tools
    "fastapi>=0.115.0, <0.119.0",                              # FastAPI framework
    "google-api-python-client>=2.157.0, <3.0.0",               # Google API client discovery
    "google-cloud-aiplatform[agent_engines]>=1.125.0, <2.0.0", # For VertexAI integrations, e.g. example store.
    "google-cloud-bigquery-storage>=2.0.0",
    "google-cloud-bigquery>=2.2.0",
    "google-cloud-bigtable>=2.32.0",                           # For Bigtable database
    "google-cloud-discoveryengine>=0.13.12, <0.14.0",          # For Discovery Engine Search Tool
    "google-cloud-secret-manager>=2.22.0, <3.0.0",             # Fetching secrets in RestAPI Tool
    "google-cloud-spanner>=3.56.0, <4.0.0",                    # For Spanner database
    "google-cloud-speech>=2.30.0, <3.0.0",                     # For Audio Transcription
    "google-cloud-storage>=2.18.0, <4.0.0",                    # For GCS Artifact service
    "google-genai>=1.45.0, <2.0.0",                            # Google GenAI SDK
    "graphviz>=0.20.2, <1.0.0",                                # Graphviz for graph rendering
    "jsonschema>=4.23.0, <5.0.0",                              # Agent Builder config validation
    "mcp>=1.10.0, <2.0.0",                                     # For MCP Toolset
    "opentelemetry-api>=1.37.0, <=1.37.0",                     # OpenTelemetry - limit upper version for sdk and api to not risk breaking changes from unstable _logs package.
    "opentelemetry-exporter-gcp-logging>=1.9.0a0, <2.0.0",
    "opentelemetry-exporter-gcp-monitoring>=1.9.0a0, <2.0.0",
    "opentelemetry-exporter-gcp-trace>=1.9.0, <2.0.0",
    "opentelemetry-exporter-otlp-proto-http>=1.36.0",
    "opentelemetry-resourcedetector-gcp>=1.9.0a0, <2.0.0",
    "opentelemetry-sdk>=1.37.0, <=1.37.0",
    "pyarrow>=14.0.0",
    "pydantic>=2.0, <3.0.0",                                   # For data validation/models
    "python-dateutil>=2.9.0.post0, <3.0.0",                    # For Vertext AI Session Service
    "python-dotenv>=1.0.0, <2.0.0",                            # To manage environment variables
    "requests>=2.32.4, <3.0.0",
    "sqlalchemy-spanner>=1.14.0",                              # Spanner database session service
    "sqlalchemy>=2.0, <3.0.0",                                 # SQL database ORM
    "starlette>=0.46.2, <1.0.0",                               # For FastAPI CLI
    "tenacity>=9.0.0, <10.0.0",                                # For Retry management
    "typing-extensions>=4.5, <5",
    "tzlocal>=5.3, <6.0",                                      # Time zone utilities
    "uvicorn>=0.34.0, <1.0.0",                                 # ASGI server for FastAPI
    "watchdog>=6.0.0, <7.0.0",                                 # For file change detection and hot reload
    "websockets>=15.0.1, <16.0.0",                             # For BaseLlmFlow
    "streamlit>0.1.0"                                                # For UI
    
]

[tool.setuptools.packages.find]
where = ["."]

[dependency-groups]
dev = [
    "pytest>=8.3.2",
    "pytest-asyncio>=0.23.7",
    "google-adk[eval]>=1.0.0",
    "google-cloud-aiplatform[adk,agent-engines,evaluation]>=1.93.0",
    "nest-asyncio>=1.6.0",
    "pandas>=2.2.3",
    "tabulate>=0.9.0",
    "agent-starter-pack>=0.14.1",
]

[tool.uv.workspace]
members = [
    "deploy_test",
]

